requirements.txt
tensorflow==2.15.0
opencv-python==4.8.0.76
numpy==1.24.3
pillow==9.5.0
streamlit==1.27.0